524c44c3 version 2.2.4

+++ src/org/thoughtcrime/securesms/util/BackupUtil.java
@@ -1,14 +1,25 @@
 package org.thoughtcrime.securesms.util;
 
 
+import android.Manifest;
 import android.content.Context;
 
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Build;
+import android.os.Environment;
+import android.text.TextUtils;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import androidx.annotation.RequiresApi;
+import androidx.documentfile.provider.DocumentFile;
 import org.thoughtcrime.securesms.R;
+import org.thoughtcrime.securesms.backup.BackupPassphrase;
 import org.thoughtcrime.securesms.database.NoExternalStorageException;
+import org.thoughtcrime.securesms.dependencies.ApplicationDependencies;
 import org.thoughtcrime.securesms.logging.Log;
+import org.thoughtcrime.securesms.permissions.Permissions;
 import org.whispersystems.libsignal.util.ByteUtil;
 
 import java.io.File;
@@ -18,10 +29,13 @@ import java.util.Calendar;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
+import java.util.Objects;
 
 public class BackupUtil {
 
   private static final String TAG = BackupUtil.class.getSimpleName();
+
+  public static final int PASSPHRASE_LENGTH = 30;
 
   public static @NonNull String getLastBackupTime(@NonNull Context context, @NonNull Locale locale) {
     try {
@@ -32,6 +46,24 @@ public class BackupUtil {
     } catch (NoExternalStorageException e) {
       Log.w(TAG, e);
       return context.getString(R.string.BackupUtil_unknown);
+    }
+  }
+
+  public static boolean isUserSelectionRequired(@NonNull Context context) {
+    return Build.VERSION.SDK_INT >= 29 && !Permissions.hasAll(context, Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE);
+  }
+
+  public static boolean canUserAccessBackupDirectory(@NonNull Context context) {
+    if (isUserSelectionRequired(context)) {
+      Uri backupDirectoryUri = TextSecurePreferences.getBackupDirectory(context);
+      if (backupDirectoryUri == null) {
+        return false;
+      }
+
+      DocumentFile backupDirectory = DocumentFile.fromTreeUri(context, backupDirectoryUri);
+      return backupDirectory != null && backupDirectory.exists() && backupDirectory.canRead() && backupDirectory.canWrite();
+    } else {
+      return Permissions.hasAll(context, Manifest.permission.WRITE_EXTERNAL_STORAGE);
     }
   }
 
@@ -69,17 +101,94 @@ public class BackupUtil {
     }
   }
 
+  public static void disableBackups(@NonNull Context context) {
+    BackupPassphrase.set(context, null);
+    TextSecurePreferences.setBackupEnabled(context, false);
+    BackupUtil.deleteAllBackups();
+
+    if (BackupUtil.isUserSelectionRequired(context)) {
+      Uri backupLocationUri = TextSecurePreferences.getBackupDirectory(context);
+
+      if (backupLocationUri == null) {
+        return;
+      }
+
+      TextSecurePreferences.clearBackupDirectory(context);
+
+      try {
+        context.getContentResolver()
+            .releasePersistableUriPermission(Objects.requireNonNull(backupLocationUri),
+                                             Intent.FLAG_GRANT_READ_URI_PERMISSION |
+                                             Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
+      } catch (SecurityException e) {
+        Log.w(TAG, "Could not release permissions", e);
+      }
+    }
+  }
+
   private static List<BackupInfo> getAllBackupsNewestFirst() throws NoExternalStorageException {
-    File             backupDirectory = StorageUtil.getBackupDirectory();
+    if (isUserSelectionRequired(ApplicationDependencies.getApplication())) {
+      return getAllBackupsNewestFirstApi29();
+    } else {
+      return getAllBackupsNewestFirstLegacy();
+    }
+  }
+
+  @RequiresApi(29)
+  private static List<BackupInfo> getAllBackupsNewestFirstApi29() {
+    Uri backupDirectoryUri = TextSecurePreferences.getBackupDirectory(ApplicationDependencies.getApplication());
+    if (backupDirectoryUri == null) {
+      Log.i(TAG, "Backup directory is not set. Returning an empty list.");
+      return Collections.emptyList();
+    }
+
+    DocumentFile backupDirectory = DocumentFile.fromTreeUri(ApplicationDependencies.getApplication(), backupDirectoryUri);
+    if (backupDirectory == null || !backupDirectory.exists() || !backupDirectory.canRead()) {
+      Log.w(TAG, "Backup directory is inaccessible. Returning an empty list.");
+      return Collections.emptyList();
+    }
+
+    DocumentFile[]   files   = backupDirectory.listFiles();
+    List<BackupInfo> backups = new ArrayList<>(files.length);
+
+    for (DocumentFile file : files) {
+      if (file.isFile() && file.getName() != null && file.getName().endsWith(".backup")) {
+        long backupTimestamp = getBackupTimestamp(file.getName());
+
+        if (backupTimestamp != -1) {
+          backups.add(new BackupInfo(backupTimestamp, file.length(), file.getUri()));
+        }
+      }
+    }
+
+    Collections.sort(backups, (a, b) -> Long.compare(b.timestamp, a.timestamp));
+
+    return backups;
+  }
+
+  public static @Nullable BackupInfo getBackupInfoFromSingleUri(@NonNull Context context, @NonNull Uri singleUri) {
+    DocumentFile documentFile = DocumentFile.fromSingleUri(context, singleUri);
+
+    if (isBackupFileReadable(documentFile)) {
+      long backupTimestamp = getBackupTimestamp(Objects.requireNonNull(documentFile.getName()));
+      return new BackupInfo(backupTimestamp, documentFile.length(), documentFile.getUri());
+    } else {
+      Log.w(TAG, "Could not load backup info.");
+      return null;
+    }
+  }
+
+  private static List<BackupInfo> getAllBackupsNewestFirstLegacy() throws NoExternalStorageException {
+    File             backupDirectory = StorageUtil.getOrCreateBackupDirectory();
     File[]           files           = backupDirectory.listFiles();
     List<BackupInfo> backups         = new ArrayList<>(files.length);
 
     for (File file : files) {
       if (file.isFile() && file.getAbsolutePath().endsWith(".backup")) {
-        long backupTimestamp = getBackupTimestamp(file);
+        long backupTimestamp = getBackupTimestamp(file.getName());
 
         if (backupTimestamp != -1) {
-          backups.add(new BackupInfo(backupTimestamp, file.length(), file));
+          backups.add(new BackupInfo(backupTimestamp, file.length(), Uri.fromFile(file)));
         }
       }
     }
@@ -96,15 +205,34 @@ public class BackupUtil {
     new SecureRandom().nextBytes(random);
 
     for (int i=0;i<30;i+=5) {
-      result[i/5] = String.format("%05d", ByteUtil.byteArray5ToLong(random, i) % 100000);
+      result[i/5] = String.format(Locale.ENGLISH,  "%05d", ByteUtil.byteArray5ToLong(random, i) % 100000);
     }
 
     return result;
   }
 
-  private static long getBackupTimestamp(File backup) {
-    String   name  = backup.getName();
-    String[] prefixSuffix = name.split("[.]");
+  public static boolean hasBackupFiles(@NonNull Context context) {
+    if (Permissions.hasAll(context, Manifest.permission.READ_EXTERNAL_STORAGE)) {
+      try {
+        File directory = StorageUtil.getBackupDirectory();
+
+        if (directory.exists() && directory.isDirectory()) {
+          File[] files = directory.listFiles();
+          return files != null && files.length > 0;
+        } else {
+          return false;
+        }
+      } catch (NoExternalStorageException e) {
+        Log.w(TAG, "Failed to read storage!", e);
+        return false;
+      }
+    } else {
+      return false;
+    }
+  }
+
+  private static long getBackupTimestamp(@NonNull String backupName) {
+    String[] prefixSuffix = backupName.split("[.]");
 
     if (prefixSuffix.length == 2) {
       String[] parts = prefixSuffix[0].split("\\-");
@@ -130,16 +258,34 @@ public class BackupUtil {
     return -1;
   }
 
+  private static boolean isBackupFileReadable(@Nullable DocumentFile documentFile) {
+    if (documentFile == null) {
+      throw new AssertionError("We do not support platforms prior to KitKat.");
+    } else if (!documentFile.exists()) {
+      Log.w(TAG, "isBackupFileReadable: The document at the specified Uri cannot be found.");
+      return false;
+    } else if (!documentFile.canRead()) {
+      Log.w(TAG, "isBackupFileReadable: The document at the specified Uri cannot be read.");
+      return false;
+    } else if (TextUtils.isEmpty(documentFile.getName()) || !documentFile.getName().endsWith(".backup")) {
+      Log.w(TAG, "isBackupFileReadable: The document at the specified Uri has an unsupported file extension.");
+      return false;
+    } else {
+      Log.i(TAG, "isBackupFileReadable: The document at the specified Uri looks like a readable backup");
+      return true;
+    }
+  }
+
   public static class BackupInfo {
 
     private final long timestamp;
     private final long size;
-    private final File file;
+    private final Uri  uri;
 
-    BackupInfo(long timestamp, long size, File file) {
+    BackupInfo(long timestamp, long size, Uri uri) {
       this.timestamp = timestamp;
       this.size      = size;
-      this.file      = file;
+      this.uri       = uri;
     }
 
     public long getTimestamp() {
@@ -150,15 +296,28 @@ public class BackupUtil {
       return size;
     }
 
-    public File getFile() {
-      return file;
+    public Uri getUri() {
+      return uri;
     }
 
     private void delete() {
-      Log.i(TAG, "Deleting: " + file.getAbsolutePath());
+      File file = new File(Objects.requireNonNull(uri.getPath()));
 
-      if (!file.delete()) {
-        Log.w(TAG, "Delete failed: " + file.getAbsolutePath());
+      if (file.exists()) {
+        Log.i(TAG, "Deleting File: " + file.getAbsolutePath());
+
+        if (!file.delete()) {
+          Log.w(TAG, "Delete failed: " + file.getAbsolutePath());
+        }
+      } else {
+        DocumentFile document = DocumentFile.fromSingleUri(ApplicationDependencies.getApplication(), uri);
+        if (document != null && document.exists()) {
+          Log.i(TAG, "Deleting DocumentFile: " + uri);
+
+          if (!document.delete()) {
+            Log.w(TAG, "Delete failed: " + uri);
+          }
+        }
       }
     }
   }
